Extract Platform enum to separate file
Moved the Platform enum from AiChatController.php to its own file
(Platform.php) to fix autoloading issues. The enum was not being
found by other classes like Executor.php because it was defined
inline within AiChatController.php.
